+++ DemoApp/ISDemoViewController.h
@@ -2,6 +2,11 @@
 
 @interface ISDemoViewController : UITableViewController
 
-@property (strong, nonatomic) NSArray *items;
+@property (nonatomic, strong) NSArray *items;
+
+@property (nonatomic, readonly) CGFloat rowHeight;
+@property (nonatomic, readonly) CGFloat topInset;
+
+- (id)initWithStyle:(UITableViewStyle)style rowHeight:(CGFloat)rowHeight topInset:(CGFloat)topInset;
 
 @end