+++ ISRefreshControl/ISScalingActivityIndicatorView.m
@@ -1,5 +1,6 @@
 #import "ISScalingActivityIndicatorView.h"
 #import "UIColor+ISRefreshControl.h"
+#import <QuartzCore/QuartzCore.h>
 
 @implementation ISScalingActivityIndicatorView
 
@@ -7,9 +8,7 @@
 {
     self = [super initWithFrame:frame];
     if (self) {
-        self.activityIndicatorViewStyle = UIActivityIndicatorViewStyleWhiteLarge;
-        self.color = [UIColor is_refreshControlColor];
-        self.transform = CGAffineTransformMakeScale(0.01f, 0.01f);
+        [self initialize];
     }
     return self;
 }
@@ -18,35 +17,53 @@
 {
     self = [super initWithCoder:coder];
     if (self) {
-        self.activityIndicatorViewStyle = UIActivityIndicatorViewStyleWhiteLarge;
-        self.color = [UIColor is_refreshControlColor];
-        self.transform = CGAffineTransformMakeScale(0.01f, 0.01f);
+        [self initialize];
     }
     return self;
+}
+
+- (void)initialize
+{
+    if ([self respondsToSelector:@selector(setColor:)]) {
+        self.activityIndicatorViewStyle = UIActivityIndicatorViewStyleWhiteLarge;
+        self.color = [UIColor is_refreshControlColor];
+        self.layer.transform = CATransform3DMakeScale(.7f, .7f, .7f);
+    } else {
+        // iOS 4.x
+        self.activityIndicatorViewStyle = UIActivityIndicatorViewStyleGray;
+    }
 }
 
 - (void)startAnimating
 {
     [super startAnimating];
     
-    CGAffineTransform rotation = CGAffineTransformMakeRotation(-M_PI_2);
-    CGAffineTransform scale = CGAffineTransformMakeScale(0.01f, 0.01f);
-    self.transform = CGAffineTransformConcat(rotation, scale);
     
-    double delayInSeconds = 0.1;
-    dispatch_time_t popTime = dispatch_time(DISPATCH_TIME_NOW, (int64_t)(delayInSeconds * NSEC_PER_SEC));
-    dispatch_after(popTime, dispatch_get_main_queue(), ^(void){
-        [UIView animateWithDuration:.2
-                         animations:^{
-                             self.transform = CGAffineTransformMakeScale(0.8f, 0.8f);
-                         }
-                         completion:^(BOOL finished) {
-                             [UIView animateWithDuration:.2
-                                              animations:^{
-                                                  self.transform = CGAffineTransformMakeScale(0.7f, 0.7f);
-                                              }];
-                         }];
-    });
+    NSArray *scaleValues;
+    if (self.activityIndicatorViewStyle == UIActivityIndicatorViewStyleWhiteLarge) {
+        scaleValues = @[@.01f, @.85f, @.7f];
+    } else {
+        scaleValues = @[@.01f, @1.2f, @1.f];
+    }
+    
+    NSTimeInterval duration = .4;
+    
+    CAKeyframeAnimation *scaleXAnimation = [CAKeyframeAnimation animationWithKeyPath:@"transform.scale.x"];
+    scaleXAnimation.duration = duration;
+    scaleXAnimation.values = scaleValues;
+    [self.layer addAnimation:scaleXAnimation forKey:@"scaleXAnimation"];
+    
+    CAKeyframeAnimation *scaleYAnimation = [CAKeyframeAnimation animationWithKeyPath:@"transform.scale.y"];
+    scaleYAnimation.duration = duration;
+    scaleYAnimation.values = scaleValues;
+    [self.layer addAnimation:scaleYAnimation forKey:@"scaleYAnimation"];
+    
+    CABasicAnimation *rotatingAnimation = [CABasicAnimation animationWithKeyPath:@"transform.rotation"];
+    rotatingAnimation.duration = duration * .8;
+    rotatingAnimation.fromValue = @(-M_PI * .8);
+    rotatingAnimation.toValue = @(.0);
+    rotatingAnimation.timingFunction = [CAMediaTimingFunction functionWithName:kCAMediaTimingFunctionEaseOut];
+    [self.layer addAnimation:rotatingAnimation forKey:@"rotatingAnimation"];
 }
 
 - (void)stopAnimating
@@ -55,15 +72,38 @@
         return;
     }
     
-    [UIView animateWithDuration:.255f
-                     animations:^{
-                         CGAffineTransform rotation = CGAffineTransformMakeRotation(M_PI_2);
-                         CGAffineTransform scale = CGAffineTransformMakeScale(0.01f, 0.01f);
-                         self.transform = CGAffineTransformConcat(rotation, scale);
-                     }
-                     completion:^(BOOL finished) {
-                         [super stopAnimating];
-                     }];
+    BOOL isOS4 = self.activityIndicatorViewStyle == UIActivityIndicatorViewStyleGray;
+    NSTimeInterval duration = .255;
+    
+    CABasicAnimation *scaleXAnimation = [CABasicAnimation animationWithKeyPath:@"transform.scale.x"];
+    scaleXAnimation.duration = duration;
+    scaleXAnimation.fromValue = isOS4 ? @1.f : @.7f;
+    scaleXAnimation.toValue = @.01f;
+    scaleXAnimation.delegate = self;
+    scaleXAnimation.fillMode = kCAFillModeForwards;
+    scaleXAnimation.removedOnCompletion = NO;
+    [self.layer addAnimation:scaleXAnimation forKey:@"scaleXAnimation"];
+    
+    CABasicAnimation *scaleYAnimation = [CABasicAnimation animationWithKeyPath:@"transform.scale.y"];
+    scaleYAnimation.duration = duration;
+    scaleYAnimation.fromValue = isOS4 ? @1.f : @.7f;
+    scaleYAnimation.toValue = @.01f;
+    scaleYAnimation.fillMode = kCAFillModeForwards;
+    scaleYAnimation.removedOnCompletion = NO;
+    [self.layer addAnimation:scaleYAnimation forKey:@"scaleYAnimation"];
+    
+    CABasicAnimation *rotatingAnimation = [CABasicAnimation animationWithKeyPath:@"transform.rotation"];
+    rotatingAnimation.duration = duration;
+    rotatingAnimation.fromValue = @(.0);
+    rotatingAnimation.toValue = @(M_PI * .3);
+    rotatingAnimation.timingFunction = [CAMediaTimingFunction functionWithName:kCAMediaTimingFunctionEaseOut];
+    [self.layer addAnimation:rotatingAnimation forKey:@"rotatingAnimation"];
+}
+
+- (void)animationDidStop:(CAAnimation *)animation finished:(BOOL)finished
+{
+    [super stopAnimating];
+    [self.layer removeAllAnimations];
 }
 
 @end